requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0

# Spectral analysis for irregular time series
tapify>=1.0.0
scipy>=1.10.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter notebooks for exploration
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Optional: Progress bars and logging
tqdm>=4.66.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0
